+++ src/common/constants/routes.constants.ts
@@ -2,6 +2,9 @@ export const ROUTE_PATH = {
   RESOURCES: 'resources',
   CREATE: 'create',
   EDIT: 'edit',
+  IMPORT: 'import',
+  PREVIEW: 'preview',
+  HUB: 'hub',
 };
 
 export const ROUTES = {
@@ -25,6 +28,10 @@ export const ROUTES = {
     uri: `/${ROUTE_PATH.RESOURCES}/external/:externalId/preview`,
     name: 'ld.externalResourcePreview',
   },
+  HUB_IMPORT_PREVIEW: {
+    uri: `/${ROUTE_PATH.IMPORT}/${ROUTE_PATH.HUB}/${ROUTE_PATH.PREVIEW}`,
+    name: 'ld.hubImportPreview',
+  },
   MANAGE_PROFILE_SETTINGS: {
     uri: `/manage-profile-settings`,
     name: 'ld.manageProfileSettings',
@@ -33,6 +40,7 @@ export const ROUTES = {
 
 export const RESOURCE_URLS = [ROUTES.RESOURCE_EDIT.uri];
 export const EXTERNAL_RESOURCE_URLS = [ROUTES.EXTERNAL_RESOURCE_PREVIEW.uri];
+export const HUB_IMPORT_URLS = [ROUTES.HUB_IMPORT_PREVIEW.uri];
 
 export const RESOURCE_EDIT_CREATE_URLS = [ROUTES.RESOURCE_EDIT.uri, ROUTES.RESOURCE_CREATE.uri];
 
@@ -46,6 +54,8 @@ export enum QueryParams {
   Type = 'type',
   Ref = 'ref',
   CloneOf = 'cloneOf',
+  Source = 'source',
+  SourceUri = 'sourceUri',
   ProfileId = 'profileId',
 }