+++ src/views/Root/components/Nav.tsx
@@ -1,6 +1,7 @@
 import { DOM_ELEMENTS } from '@/common/constants/domElementsIdentifiers.constants';
 import {
   EXTERNAL_RESOURCE_URLS,
+  HUB_IMPORT_URLS,
   MANAGE_PROFILE_SETTINGS_URLS,
   RESOURCE_EDIT_CREATE_URLS,
 } from '@/common/constants/routes.constants';
@@ -9,6 +10,7 @@ import { EditControlPane } from '@/components/EditControlPane';
 import { PreviewExternalResourcePane } from '@/components/PreviewExternalResourcePane';
 import { ViewMarcControlPane } from '@/components/ViewMarcControlPane';
 
+import { HubImportNavPane } from '@/features/hubImport';
 import { ManageProfileSettingsControlPane } from '@/features/manageProfileSettings/components/ManageProfileSettingsControlPane';
 
 import { useInputsState, useMarcPreviewState } from '@/store';
@@ -18,10 +20,15 @@ import './Nav.scss';
 export const Nav = () => {
   const isEditSectionOpen = useRoutePathPattern(RESOURCE_EDIT_CREATE_URLS);
   const isExternalResourceSectionOpen = useRoutePathPattern(EXTERNAL_RESOURCE_URLS);
+  const isHubImportOpen = useRoutePathPattern(HUB_IMPORT_URLS);
   const isManageProfileSettingsOpen = useRoutePathPattern(MANAGE_PROFILE_SETTINGS_URLS);
   const { basicValue: marcPreviewData } = useMarcPreviewState(['basicValue']);
   const { record } = useInputsState(['record']);
-  const isVisible = isEditSectionOpen || (isExternalResourceSectionOpen && record) || isManageProfileSettingsOpen;
+  const isVisible =
+    isEditSectionOpen ||
+    (isExternalResourceSectionOpen && record) ||
+    (isHubImportOpen && record) ||
+    isManageProfileSettingsOpen;
 
   return (
     isVisible && (
@@ -29,6 +36,7 @@ export const Nav = () => {
         {isEditSectionOpen && !marcPreviewData && <EditControlPane />}
         {marcPreviewData && !isExternalResourceSectionOpen && <ViewMarcControlPane />}
         {isExternalResourceSectionOpen && <PreviewExternalResourcePane />}
+        {isHubImportOpen && <HubImportNavPane />}
         {isManageProfileSettingsOpen && <ManageProfileSettingsControlPane />}
       </div>
     )