+++ CHANGELOG.md
@@ -79,6 +79,8 @@
 * Implement API schema change for Hub. Refs [UILD-723].
 * Fix incorrect link property generation when URI is not available. Fixes [UILD-720].
 * Add drag and drop profile settings editing. Refs [UILD-697].
+* Create a preview screen when importing a new Hub. Refs [UILD-725].
+* Add Continue button behavior at Hub preview screen. Refs [UILD-726].
 * Update custom settings toggle behavior. Refs [UILD-727].
 
 [UILD-552]:https://folio-org.atlassian.net/browse/UILD-552
@@ -158,6 +160,8 @@
 [UILD-723]:https://folio-org.atlassian.net/browse/UILD-723
 [UILD-720]:https://folio-org.atlassian.net/browse/UILD-720
 [UILD-697]:https://folio-org.atlassian.net/browse/UILD-697
+[UILD-725]:https://folio-org.atlassian.net/browse/UILD-725
+[UILD-726]:https://folio-org.atlassian.net/browse/UILD-726
 [UILD-727]:https://folio-org.atlassian.net/browse/UILD-727
 
 ## 1.0.5 (2025-04-30)